+++ resolwe/flow/tests/test_relations.py
@@ -6,11 +6,11 @@ from django.core.management import call_command
 
 from guardian.shortcuts import assign_perm
 from rest_framework import status
-from rest_framework.test import force_authenticate
 
 from resolwe.flow.models import Collection, Entity, Relation
-from resolwe.flow.models.entity import PositionInRelation, RelationType
+from resolwe.flow.models.entity import RelationPartition, RelationType
 from resolwe.flow.views import RelationViewSet
+from resolwe.permissions.utils import assign_contributor_permissions
 from resolwe.test import ResolweAPITestCase
 
 
@@ -27,11 +27,14 @@ class TestRelationsAPI(ResolweAPITestCase):
         call_command('loaddata', os.path.join(flow_config.path, 'tests', 'fixtures', 'relationtypes'), verbosity=0)
 
         self.collection = Collection.objects.create(name='Test collection', contributor=self.contributor)
-        collection_2 = Collection.objects.create(name='Second collection', contributor=self.contributor)
+        self.collection_2 = Collection.objects.create(name='Second collection', contributor=self.contributor)
         self.entity_1 = Entity.objects.create(name='First entity', contributor=self.contributor)
         self.entity_2 = Entity.objects.create(name='Second entity', contributor=self.contributor)
         self.entity_3 = Entity.objects.create(name='Third entity', contributor=self.contributor)
         self.entity_4 = Entity.objects.create(name='Fourth entity', contributor=self.contributor)
+
+        assign_contributor_permissions(self.collection, self.contributor)
+        assign_contributor_permissions(self.collection_2, self.contributor)
 
         rel_type_group = RelationType.objects.get(name='group')
         rel_type_series = RelationType.objects.get(name='series')
@@ -40,23 +43,30 @@ class TestRelationsAPI(ResolweAPITestCase):
             contributor=self.contributor,
             collection=self.collection,
             type=rel_type_group,
-            label='replicates',
-            name='Group relation',
+            category='replicates',
         )
-        PositionInRelation.objects.create(relation=self.relation_group, entity=self.entity_1)
-        PositionInRelation.objects.create(relation=self.relation_group, entity=self.entity_2)
+        self.group_partiton_1 = RelationPartition.objects.create(relation=self.relation_group, entity=self.entity_1)
+        self.group_partiton_2 = RelationPartition.objects.create(relation=self.relation_group, entity=self.entity_2)
 
         self.relation_series = Relation.objects.create(
             contributor=self.contributor,
-            collection=collection_2,
+            collection=self.collection_2,
             type=rel_type_series,
-            label='time-series',
-            name='Series relation',
+            category='time-series',
+            unit=Relation.UNIT_HOUR,
         )
-        PositionInRelation.objects.create(relation=self.relation_series, entity=self.entity_1, position='beginning')
-        PositionInRelation.objects.create(relation=self.relation_series, entity=self.entity_2, position='beginning')
-        PositionInRelation.objects.create(relation=self.relation_series, entity=self.entity_3, position='end')
-        PositionInRelation.objects.create(relation=self.relation_series, entity=self.entity_4, position='end')
+        self.series_partiton_1 = RelationPartition.objects.create(
+            relation=self.relation_series, entity=self.entity_1, label='beginning', position=0
+        )
+        self.series_partiton_2 = RelationPartition.objects.create(
+            relation=self.relation_series, entity=self.entity_2, label='beginning', position=0
+        )
+        self.series_partiton_3 = RelationPartition.objects.create(
+            relation=self.relation_series, entity=self.entity_3, label='end', position=1
+        )
+        self.series_partiton_4 = RelationPartition.objects.create(
+            relation=self.relation_series, entity=self.entity_4, label='end', position=1
+        )
 
         assign_perm('view_relation', self.contributor, self.relation_group)
         assign_perm('view_relation', self.contributor, self.relation_series)
@@ -73,9 +83,27 @@ class TestRelationsAPI(ResolweAPITestCase):
         resp = self._get_detail(self.relation_group.pk, user=self.contributor)
         self.assertEqual(resp.status_code, status.HTTP_200_OK)
 
-        self.assertEqual(resp.data['entities'], [
-            {'entity': self.entity_1.pk, 'position': None},
-            {'entity': self.entity_2.pk, 'position': None},
+        self.assertEqual(resp.data['collection'], self.collection.pk)
+        self.assertEqual(resp.data['type'], "group")
+        self.assertEqual(resp.data['category'], "replicates")
+        self.assertEqual(resp.data['unit'], None)
+        self.assertEqual(resp.data['partitions'], [
+            {'id': self.group_partiton_1.pk, 'entity': self.entity_1.pk, 'position': None, 'label': None},
+            {'id': self.group_partiton_2.pk, 'entity': self.entity_2.pk, 'position': None, 'label': None},
+        ])
+
+        resp = self._get_detail(self.relation_series.pk, user=self.contributor)
+        self.assertEqual(resp.status_code, status.HTTP_200_OK)
+
+        self.assertEqual(resp.data['collection'], self.collection_2.pk)
+        self.assertEqual(resp.data['type'], "series")
+        self.assertEqual(resp.data['category'], "time-series")
+        self.assertEqual(resp.data['unit'], 'hr')
+        self.assertEqual(resp.data['partitions'], [
+            {'id': self.series_partiton_1.pk, 'entity': self.entity_1.pk, 'position': 0, 'label': 'beginning'},
+            {'id': self.series_partiton_2.pk, 'entity': self.entity_2.pk, 'position': 0, 'label': 'beginning'},
+            {'id': self.series_partiton_3.pk, 'entity': self.entity_3.pk, 'position': 1, 'label': 'end'},
+            {'id': self.series_partiton_4.pk, 'entity': self.entity_4.pk, 'position': 1, 'label': 'end'},
         ])
 
     def test_filtering(self):
@@ -97,8 +125,8 @@ class TestRelationsAPI(ResolweAPITestCase):
         self.assertEqual(len(resp.data), 1)
         self.assertEqual(resp.data[0]['id'], self.relation_group.pk)
 
-        # Filtering by label
-        query_params = {'label': 'replicates'}
+        # Filtering by category
+        query_params = {'category': 'replicates'}
         resp = self._get_list(user=self.contributor, query_params=query_params)
         self.assertEqual(len(resp.data), 1)
         self.assertEqual(resp.data[0]['id'], self.relation_group.pk)
@@ -109,27 +137,31 @@ class TestRelationsAPI(ResolweAPITestCase):
         self.assertEqual(len(resp.data), 1)
         self.assertEqual(resp.data[0]['id'], self.relation_series.pk)
 
-        # Filtering by entity and position
-        query_params = {'entity': self.entity_4.pk, 'position': 'end'}
+        # Filtering by entity and label
+        query_params = {'entity': self.entity_4.pk, 'label': 'end'}
         resp = self._get_list(user=self.contributor, query_params=query_params)
         self.assertEqual(len(resp.data), 1)
         self.assertEqual(resp.data[0]['id'], self.relation_series.pk)
 
-        # Filtering by entity and position - not matching position
-        query_params = {'entity': self.entity_4.pk, 'position': 'beginning'}
+        # Filtering by entity and label - not matching label
+        query_params = {'entity': self.entity_4.pk, 'label': 'beginning'}
         resp = self._get_list(user=self.contributor, query_params=query_params)
         self.assertEqual(len(resp.data), 0)
 
-        # Filtering by position - missing entity
-        query_params = {'position': 'end'}
+        # Filtering by label - missing entity
+        query_params = {'label': 'end'}
         resp = self._get_list(user=self.contributor, query_params=query_params)
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
 
-    def test_create_without_position(self):
+    def test_create_only_entity(self):
         data = {
             'collection': self.collection.pk,
             'type': 'group',
-            'entities': [{'entity': self.entity_3.pk}, {'entity': self.entity_4.pk}],
+            'category': 'time series',
+            'partitions': [
+                {'entity': self.entity_3.pk},
+                {'entity': self.entity_4.pk},
+            ],
         }
 
         resp = self._post(data, user=self.contributor)
@@ -139,15 +171,16 @@ class TestRelationsAPI(ResolweAPITestCase):
 
         relation = Relation.objects.last()
         self.assertEqual(relation.collection, self.collection)
-        self.assertEqual(relation.positioninrelation_set.count(), 2)
-        self.assertTrue(relation.positioninrelation_set.filter(entity=self.entity_3.pk).exists())
-        self.assertTrue(relation.positioninrelation_set.filter(entity=self.entity_4.pk).exists())
+        self.assertEqual(relation.relationpartition_set.count(), 2)
+        self.assertTrue(relation.relationpartition_set.filter(entity=self.entity_3.pk).exists())
+        self.assertTrue(relation.relationpartition_set.filter(entity=self.entity_4.pk).exists())
 
     def test_create_with_position(self):
         data = {
             'collection': self.collection.pk,
             'type': 'series',
-            'entities': [
+            'category': 'time series',
+            'partitions': [
                 {'entity': self.entity_3.pk, 'position': 1},
                 {'entity': self.entity_4.pk, 'position': 2},
             ],
@@ -159,14 +192,36 @@ class TestRelationsAPI(ResolweAPITestCase):
 
         relation = Relation.objects.last()
         self.assertEqual(relation.collection, self.collection)
-        self.assertEqual(relation.positioninrelation_set.count(), 2)
-        self.assertTrue(relation.positioninrelation_set.filter(entity=self.entity_3.pk, position=1).exists())
-        self.assertTrue(relation.positioninrelation_set.filter(entity=self.entity_4.pk, position=2).exists())
+        self.assertEqual(relation.relationpartition_set.count(), 2)
+        self.assertTrue(relation.relationpartition_set.filter(entity=self.entity_3.pk, position=1).exists())
+        self.assertTrue(relation.relationpartition_set.filter(entity=self.entity_4.pk, position=2).exists())
+
+    def test_create_with_label(self):
+        data = {
+            'collection': self.collection.pk,
+            'type': 'series',
+            'category': 'time series',
+            'partitions': [
+                {'entity': self.entity_3.pk, 'label': 'Hr01'},
+                {'entity': self.entity_4.pk, 'label': 'Hr02'},
+            ],
+        }
+        resp = self._post(data, user=self.contributor)
+        self.assertEqual(resp.status_code, status.HTTP_201_CREATED)
+
+        self.assertEqual(Relation.objects.count(), 3)
+
+        relation = Relation.objects.last()
+        self.assertEqual(relation.collection, self.collection)
+        self.assertEqual(relation.relationpartition_set.count(), 2)
+        self.assertTrue(relation.relationpartition_set.filter(entity=self.entity_3.pk, label='Hr01').exists())
+        self.assertTrue(relation.relationpartition_set.filter(entity=self.entity_4.pk, label='Hr02').exists())
 
     def test_create_with_missing_type(self):
         data = {
             'collection': self.collection.pk,
-            'entities': [
+            'category': 'time series',
+            'partitions': [
                 {'entity': self.entity_3.pk, 'position': 1},
                 {'entity': self.entity_4.pk, 'position': 2},
             ],
@@ -174,10 +229,37 @@ class TestRelationsAPI(ResolweAPITestCase):
         resp = self._post(data, user=self.contributor)
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
 
+    def test_create_missing_category(self):
+        data = {
+            'collection': self.collection.pk,
+            'type': 'series',
+            'partitions': [
+                {'entity': self.entity_3.pk, 'position': 1},
+                {'entity': self.entity_4.pk, 'position': 2},
+            ],
+        }
+        resp = self._post(data, user=self.contributor)
+        self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_create_duplicated_category(self):
+        data = {
+            'collection': self.collection.pk,
+            'type': 'group',
+            'category': 'RePlIcAtEs',
+            'partitions': [
+                {'entity': self.entity_3.pk},
+                {'entity': self.entity_4.pk},
+            ],
+        }
+
+        resp = self._post(data, user=self.contributor)
+        self.assertEqual(resp.status_code, status.HTTP_409_CONFLICT)
+
     def test_create_missing_collection(self):
         data = {
             'type': 'series',
-            'entities': [
+            'category': 'time series',
+            'partitions': [
                 {'entity': self.entity_3.pk, 'position': 1},
                 {'entity': self.entity_4.pk, 'position': 2},
             ],
@@ -186,76 +268,76 @@ class TestRelationsAPI(ResolweAPITestCase):
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
 
     def test_update(self):
-        new_collection = Collection.objects.create(name='Test collection 2', contributor=self.contributor)
-        data = {'collection': new_collection.pk}
+        data = {'collection': self.collection_2.pk}
         resp = self._patch(self.relation_group.pk, data, user=self.contributor)
         self.assertEqual(resp.status_code, status.HTTP_200_OK)
 
         self.relation_group.refresh_from_db()
-        self.assertEqual(self.relation_group.collection, new_collection)
+        self.assertEqual(self.relation_group.collection, self.collection_2)
 
         # Relation type cannot be changed
         data = {'type': 'series'}
         resp = self._patch(self.relation_group.pk, data, user=self.contributor)
         self.assertEqual(resp.status_code, status.HTTP_204_NO_CONTENT)
 
-        # Entities nust be updated through special endpoints
         data = {
-            'entities': [
-                {'entity': self.entity_3.pk, 'position': 1},
+            'partitions': [
+                {'entity': self.entity_3.pk},
+                {'entity': self.entity_4.pk},
             ],
         }
         resp = self._patch(self.relation_group.pk, data, user=self.contributor)
-        self.assertEqual(resp.status_code, status.HTTP_204_NO_CONTENT)
+        self.assertEqual(resp.status_code, status.HTTP_200_OK)
+        self.assertEqual(self.relation_group.entities.count(), 2)
+        self.assertTrue(self.relation_group.relationpartition_set.filter(entity=self.entity_3.pk).exists())
+        self.assertTrue(self.relation_group.relationpartition_set.filter(entity=self.entity_4.pk).exists())
+
+    def test_update_superuser(self):
+        data = {'collection': self.collection_2.pk}
+        resp = self._patch(self.relation_group.pk, data, user=self.admin)
+        self.assertEqual(resp.status_code, status.HTTP_200_OK)
+        self.relation_group.refresh_from_db()
+        self.assertEqual(self.relation_group.collection, self.collection_2)
+
+    def test_update_different_user(self):
+        data = {'collection': self.collection_2.pk}
+        resp = self._patch(self.relation_group.pk, data, user=self.user)
+        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
+        self.relation_group.refresh_from_db()
+        self.assertEqual(self.relation_group.collection, self.collection)
+
+        assign_contributor_permissions(self.collection, self.user)
+
+        resp = self._patch(self.relation_group.pk, data, user=self.user)
+        self.assertEqual(resp.status_code, status.HTTP_200_OK)
+        self.relation_group.refresh_from_db()
+        self.assertEqual(self.relation_group.collection, self.collection_2)
 
     def test_delete(self):
         resp = self._delete(self.relation_group.pk, user=self.contributor)
         self.assertEqual(resp.status_code, status.HTTP_204_NO_CONTENT)
+        self.assertFalse(Relation.objects.filter(pk=self.relation_group.pk).exists())
 
-        self.assertEqual(Relation.objects.count(), 1)
+    def test_delete_superuser(self):
+        resp = self._delete(self.relation_group.pk, user=self.admin)
+        self.assertEqual(resp.status_code, status.HTTP_204_NO_CONTENT)
+        self.assertFalse(Relation.objects.filter(pk=self.relation_group.pk).exists())
 
-    def test_add_entity(self):
-        data = [{'entity': self.entity_3.pk}]
+    def test_delete_different_user(self):
+        resp = self._delete(self.relation_group.pk, user=self.user)
+        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
+        self.assertTrue(Relation.objects.filter(pk=self.relation_group.pk).exists())
 
-        request = self.factory.post('', data=data, format='json')
-        force_authenticate(request, self.contributor)
-        self.add_entity_viewset(request, pk=self.relation_group.pk)
+        assign_contributor_permissions(self.collection, self.user)
 
-        self.assertEqual(self.relation_group.positioninrelation_set.count(), 3)
-        self.assertEqual(self.relation_group.positioninrelation_set.last().position, None)
+        resp = self._delete(self.relation_group.pk, user=self.contributor)
+        self.assertEqual(resp.status_code, status.HTTP_204_NO_CONTENT)
+        self.assertFalse(Relation.objects.filter(pk=self.relation_group.pk).exists())
 
-    def test_add_entity_with_position(self):
-        data = [{'entity': self.entity_3.pk, 'position': 'latest'}]
+    def test_delete_collection(self):
+        self.collection.delete()
+        self.assertFalse(Relation.objects.filter(pk=self.relation_group.pk).exists())
 
-        request = self.factory.post('', data=data, format='json')
-        force_authenticate(request, self.contributor)
-        self.add_entity_viewset(request, pk=self.relation_group.pk)
-
-        self.assertEqual(self.relation_group.positioninrelation_set.count(), 3)
-        self.assertEqual(self.relation_group.positioninrelation_set.last().position, 'latest')
-
-    def test_add_entity_position_none(self):
-        data = [{'entity': self.entity_3.pk, 'position': None}]
-
-        request = self.factory.post('', data=data, format='json')
-        force_authenticate(request, self.contributor)
-        self.add_entity_viewset(request, pk=self.relation_group.pk)
-
-        self.assertEqual(self.relation_group.positioninrelation_set.count(), 3)
-        self.assertEqual(self.relation_group.positioninrelation_set.last().position, None)
-
-    def test_add_entity_missing_key(self):
-        data = [{'foo': self.entity_3.pk}]
-        request = self.factory.post('', data=data, format='json')
-        force_authenticate(request, self.contributor)
-        self.add_entity_viewset(request, pk=self.relation_group.pk)
-
-        self.assertEqual(self.relation_group.positioninrelation_set.count(), 2)
-
-    def test_remove_entity(self):
-        data = {'ids': [self.entity_2.pk]}
-        request = self.factory.post('', data=data, format='json')
-        force_authenticate(request, self.contributor)
-        self.remove_entity_viewset(request, pk=self.relation_group.pk)
-
-        self.assertEqual(self.relation_group.positioninrelation_set.count(), 1)
+    def test_delete_entity(self):
+        self.entity_1.delete()
+        self.assertEqual(self.relation_group.entities.count(), 1)